problem was drawer WIP also deltaTime maybe

+++ WindowsProject1/Time.h
@@ -1,9 +1,11 @@
+#include "Windows.h"
+
 #pragma once
 class Time {
 public:
 	static double deltaTime;
-	static double currentTime;
-	static double lastTime;
+	static LONGLONG currentTime;
+	static LONGLONG lastTime;
 
 	static void calcTimes();
 };
